+++ app/Http/Controllers/MembersController.php
@@ -6,8 +6,10 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+use App\Http\Helpers\ImageHelper;
 use App\Http\Requests\MemberRequest;
 use File;
+use Session;
 
 use App\Member;
 
@@ -15,6 +17,8 @@ class MembersController extends Controller
 {
 
     private $member;
+
+    private $imageHelper;
 
     public function __construct()
     {
@@ -24,7 +28,9 @@ class MembersController extends Controller
         //Check if this user is a admin
         $this->middleware('admin');
 
-        $this->member = new Member();  
+        $this->member = new Member();
+
+        $this->imageHelper = new ImageHelper();
     }
 
     /**
@@ -35,10 +41,26 @@ class MembersController extends Controller
     public function index()
     {
         //Return all the results
-        $members = $this->member->orderBy('type', 'asc')->get();
+        $members = $this->member->orderBy('type', 'asc')->orderBy('order', 'asc')->get();
 
         //Passing the Role to use its constants
         return view('admin.member.index', compact('members'));
+    }
+
+    public function staff()
+    {
+        //Find the staff members
+        $members = $this->member->where('type', Member::STAFF)->orderBy('order', 'asc')->get();
+
+        return view('pages.about.staff', compact('members'));
+    }
+
+    public function boardOfdirector()
+    {
+        //Find the board directors members
+        $members = $this->member->where('type', Member::BOARD_DIRECTOR)->orderBy('order', 'asc')->get();
+
+        return view('pages.about.boardofdirectors', compact('members'));
     }
 
     /**
@@ -64,13 +86,27 @@ class MembersController extends Controller
         $type = $request->input('type');
         $order = $request->input('order');
 
-        $this->adjustImagesOrder($type, $order);
+        /* 
+            Laravel can not handle using the request manager the validation
+            of the size image that exceeds the configuration of php.in
+        */
+        if($this->imageHelper->fileImageSizeExceeded($request->file('image'))){
+            //Set the message and the error class
+            Session::flash('message', 'The file image can not be greater than 2MB!'); 
+            Session::flash('alert-class', 'alert-danger');
+
+            return redirect()->back()
+                             ->withInput()
+                             ->withErrors('image');
+        }
+
+        $this->imageHelper->adjustOrder($type, $order, $this->member, 'store');
 
         //Creating the new member
         $member = $this->member->create($request->all());
 
         //Moves and sets the uploaded image
-        $this->uploadImage($request, $member); 
+        $this->imageHelper->uploadImage($request, $member); 
 
         //Saving
         $member->save();
@@ -115,20 +151,37 @@ class MembersController extends Controller
         $type = $request->input('type');
         $order = $request->input('order');
 
-        $this->adjustImagesOrder($type, $order);
+        /* 
+            Laravel can not handle using the request manager the validation
+            of the size image that exceeds the configuration of php.in
+        */
+        if($this->imageHelper->fileImageSizeExceeded($request->file('image'))){
+            //Set the message and the error class
+            Session::flash('message', 'The file image can not be greater than 2MB!'); 
+            Session::flash('alert-class', 'alert-danger');
 
-        //Fill the member with the new form fields
+            return redirect()->back();
+        }
+
+        //Find the member
         $member = $this->member->find($id);
-        $member->fill($request->all());
+       
+        //If the admin changed the member order, so it is necessary adjust the others members.
+        if($member->order != $request->input('order')){
+            $this->imageHelper->adjustOrder($type, $order, $this->member, 'update');
+        }
+
+        //Fill this member with the new form information
+        $member->fill($request->all());        
         
         //Checking if there was uploaded a image
         if ($request->file('image') != null) {
-            
+
             //Delete the old member image from the filesystem
             File::delete($member->path);
 
             //Moves and sets the new uploaded image
-            $this->uploadImage($request, $member);            
+            $this->imageHelper->uploadImage($request, $member);            
         }        
         
         $member->save();
@@ -147,63 +200,17 @@ class MembersController extends Controller
     public function destroy($id)
     {
         $member = $this->member->find($id);
-        
 
-        //Moving the order of all images of sted backward
-        $members = $member->where('type', $member->type)->where('order','>', $member->order)->get();            
-        foreach ($members as $member) {
-
-            $member->order = $member->order - 1;
-            $member->save();
-        }            
+        //Adjust the order of the members when delete this .
+        $this->imageHelper->adjustOrder($member->type, $member->order, $member, 'destroy');
 
         //Delete the member image on the filesystem
         File::delete($member->path);
 
         //Delete the image from the database
-        $member->delete();
-        
+        $member->delete();        
         
         //Sending the user to the accounts page
         return redirect()->route('member/index');
-    }
-
-    /*
-    Move all images' order one step forward if the parameterized is already used
-    */
-    public function adjustImagesOrder($type, $order){
-        $members = Member::where('type', $type)->where('order','>=', $order)->get();  
-
-        foreach ($members as $member) {
-
-            $member->order = $member->order + 1;
-            $member->save();
-        }
-    }
-
-    /*
-    Gets the image uploaded from the form, and moves it to the directory.
-    */
-    public function uploadImage($request, $image){
-        //Getting the upload image
-        
-        //There is a erro on the upload image, show message error
-        if($request->file('image')->getError() == 1 && $request->file('image')->getClientSize() == 0){
-            
-            //Set the message and the error class
-            Session::flash('message', 'The can not exceed 2MB!'); 
-            Session::flash('alert-class', 'alert-danger');
-        } else {
-            dd(false);
-        }
-
-        $imageName = $request->input('name') . '.' . $request->file('image')->getClientOriginalExtension();
-       
-        //Moving the image to the folder, and getting the path
-        $request->file('image')->move(base_path() . '/public/' . Member::DIRECTORY, $imageName);
-        $imagePath = Member::DIRECTORY . '/' . $imageName;
-
-        //Setting the path
-        $image->path = $imagePath;
-    }
+    }    
 }
